Extending the task model to incorporate more discrete tasks
The original script defined a few operations that comprised a set of steps, like updating
the requirements file. This takes the project in a slightly different direction by defining
discrete tasks: Clone, Checkout, New Branch, Merge. This approach (hopefully) creates a more
flexible tool to operate on multiple repositories for a variety of tasks.

@todo - remove `hub` dependency in favor of a straight rest api call.
@todo - update the README file with clearer instructions.